back-end/backend/python-ai-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.0.0